+++ RiotCaller.AspNetCore/Models/LolApiRequest.cs
@@ -10,6 +10,7 @@ using System.Net.Http;
 using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 using RiotGamesApi.AspNetCore.RateLimit;
+using RiotGamesApi.AspNetCore.RateLimit.Property;
 using RiotGamesApi.AspNetCore.RiotApi.Enums;
 using RiotGamesApi.AspNetCore.RiotApi.TournamentEndPoints;
