+++ RiotCaller.AspNetCore/RateLimit/ApiLimit.cs
@@ -1,7 +1,5 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-using RiotGamesApi.AspNetCore.Enums;
+﻿using RiotGamesApi.AspNetCore.Enums;
+using System;
 
 namespace RiotGamesApi.AspNetCore.RateLimit
 {
@@ -14,10 +12,10 @@ namespace RiotGamesApi.AspNetCore.RateLimit
             LimitType = limitType;
         }
 
-        public RateLimitType LimitType { get; internal set; }
         public DateTime ChainStartTime { get; internal set; }
         public int Counter { get; internal set; }
         public int Limit { get; internal set; }
+        public RateLimitType LimitType { get; internal set; }
         public TimeSpan Time { get; internal set; }
 
         public override string ToString()
