+++ RiotCaller.AspNetCore/RateLimit/Property/RateLimitProperties.cs
@@ -1,0 +1,16 @@
+﻿using RiotGamesApi.AspNetCore.Enums;
+
+namespace RiotGamesApi.AspNetCore.RateLimit.Property
+{
+    public class RateLimitProperties
+    {
+        public LolApiName ApiName { get; set; }
+        public string Platform { get; set; }
+        public LolUrlType UrlType { get; set; }
+
+        public override string ToString()
+        {
+            return $"{Platform}:{UrlType}:{ApiName}";
+        }
+    }
+}
